Add files via upload
This is Part 1 and 2 of ESC Project 190. I have rewritten part 1 and finished part 2. I have tested both and concluded both are working as intended. The test infrastructure is still present inside this file. It must be removed before final submissions, as the final assignment should not have any print statements within.