Add files via upload
This file contains a working version of Parts 1, 2, and 3 for the first ESC190 project "Autocomplete". It contains testing infrastructure built in. This code passes all tests for wiktionary.txt. However, there were memory issues when testing on cities.txt. This is a known issue and is looking to be addressed soon.